Further cleanup. (idle.c, viconfig.c)
