+++ src/boot/idle.c
@@ -2,10 +2,8 @@
 #include "global.h"
 #include "vt.h"
 #include "sched.h"
-#include "padmgr.h"
 #include "ultra64/vi.h"
 #include "ultra64/pi.h"
-//#include "def/game.h"
 #include "def/gettime.h"
 #include "def/heaps.h"
 #include "def/idle.h"
